+++ tests/testthat/test-centrality_description.R
@@ -1,5 +1,5 @@
 test_that(
-  desc = "contingency_table works",
+  desc = "centrality description works as expected",
   code = {
     options(tibble.width = Inf)
 
@@ -14,11 +14,12 @@ test_that(
         k = list(2L, 3L, 3L, 2L),
         conf.level = list(0.89, 0.99, 0.90, 0.50)
       ),
-      .f = statsExpressions::centrality_description
+      .f = centrality_description
     )
 
     set.seed(123)
-    expect_snapshot(df)
+    expect_snapshot(select(df, -expression))
+    expect_snapshot(df$expression)
 
     # data with NAs
     set.seed(123)
@@ -31,10 +32,28 @@ test_that(
         k = list(2L, 3L, 3L, 2L),
         conf.level = list(0.89, 0.99, 0.90, 0.50)
       ),
-      .f = statsExpressions::centrality_description
+      .f = centrality_description
     )
 
     set.seed(123)
-    expect_snapshot(df_na)
+    expect_snapshot(select(df_na, -expression))
+    expect_snapshot(df_na$expression)
+  }
+)
+
+
+test_that(
+  desc = "centrality description works when variable is named `variable`",
+  code = {
+    options(tibble.width = Inf)
+
+    df_var <- dplyr::rename(mtcars, variable = cyl)
+
+    set.seed(123)
+    res <- suppressWarnings(centrality_description(df_var, variable, wt))
+
+    set.seed(123)
+    expect_snapshot(select(res, -expression))
+    expect_snapshot(res$expression)
   }
 )