+++ tests/testthat/test-two_sample_bayes.R
@@ -1,9 +1,9 @@
 test_that(
-  desc = "bayes factor (independent samples t-test)",
+  desc = "bayesian (independent samples t-test)",
   code = {
     skip_if(getRversion() < "4.0")
 
-    # bayes factor (independent samples t-test) ----------------------
+    # bayesian (independent samples t-test) ----------------------
 
     # expression
     set.seed(123)
@@ -18,12 +18,12 @@ test_that(
       )
 
     # call
-    expect_snapshot(as.character(df$expression[[1]]))
+    expect_snapshot(df$expression)
   }
 )
 
 test_that(
-  desc = "bayes factor (paired t-test)",
+  desc = "bayesian (paired t-test)",
   code = {
 
     # Bayes factor (paired t-test) ---------------------------------------------
@@ -48,7 +48,7 @@ test_that(
         bf.prior = 0.8,
       )
 
-    # check bayes factor values
+    # check bayesian values
     expect_equal(df_results$bf10[[1]], 40.36079, tolerance = 0.001)
 
     # expression
@@ -60,12 +60,11 @@ test_that(
         x = "condition",
         y = desire,
         paired = TRUE,
-        bf.prior = 0.8,
-        top.text = "bla is ulalala"
+        bf.prior = 0.8
       )
 
     # call
-    expect_snapshot(as.character(df$expression[[1]]))
+    expect_snapshot(df$expression)
   }
 )