+++ tests/testthat/test-z_meta_random_bayes.R
@@ -8,6 +8,8 @@ test_that(
     skip_if_not_installed("metaBMA")
     skip_if(getRversion() < "4.0")
 
+    options(tibble.width = Inf)
+
     # setup
     set.seed(123)
 
@@ -34,21 +36,19 @@ test_that(
       class = c("tbl_df", "tbl", "data.frame")
     )
 
-    # getting bayes factor in favor of null hypothesis
+    # getting bayesian in favor of null hypothesis
     set.seed(123)
     df <- suppressWarnings(meta_analysis(
       type = "bayes",
       data = df1,
       k = 3,
       iter = 1000,
-      summarize = "integrate",
-      top.text = "ayyo arecha"
+      summarize = "integrate"
     ))
 
     expect_type(df, "list")
-    expect_identical(class(df), c("tbl_df", "tbl", "data.frame"))
 
     expect_snapshot(dplyr::select(df, -expression))
-    expect_snapshot(as.character(df$expression[[1]]))
+    expect_snapshot(df$expression)
   }
 )